streamlit/requirements.txt
streamlit
requests
minio
faiss-cpu
numpy
opencv-python>=4.10.0
